Switch from dummy data to database data
Modify the DMQ for displaying troopers table (with JOINs) to not use
whitespace in aliases. Avoid whitespace because it does not work well
with JS attributes. Update the query in troopers.js getTroopers() to
reflect this change as well.

Change the <table> in the troopers template to use the query results
instead of dummy data.